+++ handlers/aws/utils.py
@@ -2,7 +2,6 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 import os
-from functools import lru_cache
 from typing import Any, Callable, Optional
 
 import boto3
@@ -30,6 +29,27 @@ CONFIG_FROM_PAYLOAD: str = "CONFIG_FROM_PAYLOAD"
 CONFIG_FROM_S3FILE: str = "CONFIG_FROM_S3FILE"
 
 INTEGRATION_SCOPE_GENERIC: str = "generic"
+
+
+def get_lambda_region() -> str:
+    """
+    Get the AWS region where the Lambda function is running.
+
+    Returns the value of the `AWS_REGION` environment variable. If the
+    `AWS_REGION` variable is not set, it returns the value of the
+    `AWS_DEFAULT_REGION` variable.
+
+    If neither variable is set, it raises a `ValueError`.
+
+    Returns:
+        str: The AWS region.
+    """
+    region = os.getenv("AWS_REGION") or os.getenv("AWS_DEFAULT_REGION")
+
+    if region is None:
+        raise ValueError("AWS region not found in environment variables.")
+
+    return region
 
 
 def get_sqs_client() -> BotoBaseClient:
@@ -386,49 +406,34 @@ def get_account_id_from_arn(lambda_arn: str) -> str:
     return arn_components[4]
 
 
-@lru_cache()
-def describe_regions(all_regions: bool = True) -> Any:
-    """
-    Fetches all regions from AWS and returns the response.
-
-    :return: The response from the describe_regions method
-    """
-    return get_ec2_client().describe_regions(AllRegions=all_regions)
-
-
 def get_input_from_log_group_subscription_data(
-    config: Config, account_id: str, log_group_name: str, log_stream_name: str
+    config: Config, account_id: str, log_group_name: str, log_stream_name: str, region: str
 ) -> tuple[str, Optional[Input]]:
     """
-    This function is not less resilient than the previous get_log_group_arn_and_region_from_log_group_name()
-    We avoid to call the describe_log_streams on the logs' client, since we have no way to apply the proper
-    throttling because we'd need to know the number of concurrent lambda running at the time of the call.
-    In order to not hardcode the list of regions we rely on ec2 DescribeRegions - as much weird as it is - that I found
-    no information about having any kind of throttling. We add IAM permissions for it in deployment.
+    Look up for the input in the configuration using the information
+    from the log event.
+
+    It looks for the log stream arn, if not found it looks for the
+    log group arn.
     """
-    all_regions = describe_regions(all_regions=True)
-    assert "Regions" in all_regions
-    for region_data in all_regions["Regions"]:
-        region = region_data["RegionName"]
+    partition = "aws"
+    if "gov" in region:
+        partition = "aws-us-gov"
 
-        aws_or_gov = "aws"
-        if "gov" in region:
-            aws_or_gov = "aws-us-gov"
+    log_stream_arn = (
+        f"arn:{partition}:logs:{region}:{account_id}:log-group:{log_group_name}:log-stream:{log_stream_name}"
+    )
+    event_input = config.get_input_by_id(log_stream_arn)
 
-        log_stream_arn = (
-            f"arn:{aws_or_gov}:logs:{region}:{account_id}:log-group:{log_group_name}:log-stream:{log_stream_name}"
-        )
-        event_input = config.get_input_by_id(log_stream_arn)
+    if event_input is not None:
+        return log_stream_arn, event_input
 
-        if event_input is not None:
-            return log_stream_arn, event_input
+    log_group_arn_components = log_stream_arn.split(":")
+    log_group_arn = f"{':'.join(log_group_arn_components[:-2])}:*"
+    event_input = config.get_input_by_id(log_group_arn)
 
-        log_group_arn_components = log_stream_arn.split(":")
-        log_group_arn = f"{':'.join(log_group_arn_components[:-2])}:*"
-        event_input = config.get_input_by_id(log_group_arn)
-
-        if event_input is not None:
-            return log_group_arn, event_input
+    if event_input is not None:
+        return log_group_arn, event_input
 
     return f"arn:aws:logs:%AWS_REGION%:{account_id}:log-group:{log_group_name}:*", None