+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+### v1.17.2 - 2024/09/24
+##### Bug fixes
+* Remove call to EC2:DescribeRegions API in the cloudwatch-logs input [811](https://github.com/elastic/elastic-serverless-forwarder/pull/811).
+
 ### v1.17.1 - 2024/09/23
 ##### Bug fixes
-* Cache EC2:DescribeRegion API response to avoid throttling and improve performance [803](https://github.com/elastic/elastic-serverless-forwarder/pull/803).
+* Cache EC2:DescribeRegions API response to avoid throttling and improve performance [803](https://github.com/elastic/elastic-serverless-forwarder/pull/803).
 
 ### v1.17.0 - 2024/07/10
 ##### Features